+++ lua/obsidian/img_paste.lua
@@ -115,7 +115,9 @@ M.paste_img = function(opts)
       fname = opts.default_name
     else
       fname = util.input("Enter file name: ", { default = opts.default_name, completion = "file" })
-      if not fname then
+      if fname == "" then
+        fname = opts.default_name
+      elseif not fname then
         log.warn "Paste aborted"
         return
       end
@@ -125,7 +127,10 @@ M.paste_img = function(opts)
   assert(fname)
   fname = util.strip_whitespace(fname)
 
-  if fname == "" then
+  -- Verify filename
+  if util.contains_invalid_characters(fname) then
+    log.warn "Links will not work with file names containing any of these characters in Obsidian: # ^ [ ] |"
+  elseif fname == "" then
     log.err "Invalid file name"
     return
   end