+++ lua/obsidian/util.lua
@@ -1374,6 +1374,16 @@ util.buffer_is_empty = function(bufnr)
   end
 end
 
+--- Check if a string contains invalid characters.
+---
+--- @param fname string
+---
+--- @return boolean
+util.contains_invalid_characters = function(fname)
+  local invalid_chars = "#^%[%]|"
+  return string.find(fname, "[" .. invalid_chars .. "]") ~= nil
+end
+
 ---Check if a string is NaN
 ---
 ---@param v any